+++ packages/terra-worklist-data-grid/src/proptypes/rowShape.js
@@ -6,10 +6,7 @@ const rowShape = PropTypes.shape({
    * An identifier to uniquely identify the row within the grid.
    */
   id: PropTypes.string.isRequired,
-  /**
-   * String that specifies the row height. Any valid CSS height value is accepted.
-   */
-  height: PropTypes.string,
+
   /**
    * Data to be displayed in the cells of the row. Cells will be rendered in the row in the order given.
    */