+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Updated examples and tests for `terra-worklist-data-grid` to remove row height
 
 ## 1.23.0 - (June 12, 2023)